backend/requirements.txt
Django==4.2.14
django-filter
djangorestframework==3.15.2
djoser==2.2.3
reportlab==4.2.2
chardet==5.2.0
django-cors-headers
reportlab==4.2.2
pytest==6.2.4
pytest-django==4.4.0
python-dotenv>=0.14
pytest-pythonpath==0.7.3
PyYAML==6.0
psycopg2-binary==2.9.3 